Version 0.2
This is the first public release, with some important bug fixes.

The routines that copy DeltaTalk data files to the NVDA program folder have been corrected so that administrative access is only requested when necessary. This eliminates the need to run NVDA as an administrator when installing the add-on.

The "installTasks.py" file now supports internationalization to maintain consistency with the main synthesizer code.

More log messages have been added to the main synthesizer code to make it easier to debug and identify possible problems.

The documentation for the add-on (which was previously just an initial draft) has been rewritten and updated.

The old codes were removed from the add-on because they didn't work and were obsolete.

Translations into Brazilian and european Portuguese have been added for the add-on messages.