Version 0.3
A logic has been implemented that checks the synthesizer's internal pronunciation dictionary (Brport.lng) and automatically copies it to the NVDA program folder if changes to the original file included in the add-on package are detected.
An integrated symbol dictionary has been included for DeltaTalk, allowing it to interpret punctuation marks in it's own way.
The add-on now uses "log" (imported from "logHandler") instead of "logging", for better integration with NVDA.
Experimental support has been included for audio playback using the "nvwave" system, with audio generation in multiple blocks and asynchronous playback.

  This activates initial support for secondary audio devices and audio ducking mode (Shift+NVDA+D).
  This functionality is still disabled by default and can be enabled for testing by changing the line "self._use_nvwave = False" to "True" in the code of the add-on.

Routines have been implemented to remove the DeltaTalk data files from the NVDA program folder if the add-on is uninstalled. Note that administrator privileges may be required.