Version 0.4
The functionality of the old "Pausing Information" add-on was integrated into DeltaTalk as a global plugin called "Virtual Vision Mode".
The current version is identical to the last version of the original add-on, but will be updated regularly.
The installTasks.py file now includes a routine that checks for the presence of the old add-on and removes it if it is installed.
A new settings panel has been created for DeltaTalk and added to the NVDA settings dialog. This panel will be expanded with new configuration options over time.
The add-on now works completely independently, eliminating the need to copy the DeltaTalk data files to the NVDA program folder.
The corresponding routines for copying and removing these files have been removed from the main code of add-on and from the installTasks.py file.
This should solve most of the problems reported by some users, but we cannot guarantee that they will actually be solved.